pyproject.toml
[tool.poetry]
name = "media_management_api"
version = "3.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "~=3.2.0"
psycopg2-binary = "2.8.4"
PyJWT = "~=2.1.0"
redis = "2.10.3"
django-redis-cache = "~=3.0.0"
hiredis = "~=2.0.0"
djangorestframework = "~=3.12.0"
djangorestframework-csv = "2.1.0"
Markdown = "~=3.3.0"
django-filter = "1.1.0"
django-cors-headers = "~=3.8.0"
boto = "2.48.0"
Pillow = "9.0.1"
python-magic = "0.4.15"
requests = "2.28.1"
pep8 = "1.7.1"
boto3 = "1.24.34"
IIIFingest = "1.0.5"

[tool.poetry.dev-dependencies]
coverage = "5.5"
coverage-badge = "1.0.1"
django-debug-toolbar = "1.3.2"
flake8 = "^5.0.4"
mock = "1.3.0"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
